Intermediary commit. Refactoring things to be clearer.

+++ arbre.h
@@ -35,22 +35,22 @@ extern "C" {
 #define MAX_FACTORS 30
 
 /*typedef struct t_element {
-	long long unsigned elt;
+	int unsigned elt;
 } t_element;*/
 
-typedef long long unsigned t_element;
+typedef int unsigned t_element;
 
 typedef struct t_noeud {
 	//t_element valeur;
 	t_element cle;
 	t_element* valeur;
-	long long unsigned val_size;
+	int unsigned val_size;
 	unsigned hauteur;
 	struct t_noeud *gauche;
 	struct t_noeud *droit;
 } t_arbre;
 
-t_arbre *inserer_arbre(t_arbre *a, t_element v, t_element* factors, long long unsigned v_size); // Déclaration, utile pour la suite, la définitione est plus bas.
+t_arbre *inserer_arbre(t_arbre *a, t_element v, t_element* factors, int unsigned v_size); // Déclaration, utile pour la suite, la définitione est plus bas.
 t_arbre *equilibrer_arbre(t_arbre *a);
 t_arbre *creer_arbre(t_element e, t_element* valeur, t_arbre *g, t_arbre *d);
